update local changes

+++ src/entity/Salary.java
@@ -74,7 +74,7 @@ public class Salary {
         System.out.printf("Employee ID      : %s%n", employeeId);
         System.out.printf("Period           : %02d/%d%n", month, year);
         System.out.printf("Basic Salary     : %,.0f VND%n", basicSalary);
-        System.out.printf("Overtime Pay     : %,.0f VND  (%d hrs)%n", overtimePay, overtimeHours);
+        System.out.printf("Overtime Pay     : %,.0f VND  (%f hrs)%n", overtimePay, overtimeHours);
         System.out.printf("Absence Deduction: %,.0f VND  (%d days)%n", absenceDeduction, absentDays);
         System.out.printf("Working Days     : %d%n", workingDays);
         System.out.println("--------------------------------------------");
@@ -104,7 +104,6 @@ public class Salary {
                 Double.parseDouble(p[5]),
                 Integer.parseInt(p[6]),
                 Integer.parseInt(p[7]),
-                Integer.parseInt(p[8]));
+                Double.parseDouble(p[8]));
     }
-
 }
